+++ app/Models/EventType.php
@@ -5,6 +5,21 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Model;
 use Spatie\Activitylog\Traits\LogsActivity;
 
+/**
+ * App\Models\EventType
+ *
+ * @property int $id
+ * @property string $name
+ * @property-read \Illuminate\Database\Eloquent\Collection|\Spatie\Activitylog\Models\Activity[] $activities
+ * @property-read int|null $activities_count
+ * @property-read \App\Models\Event $events
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\EventType newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\EventType newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\EventType query()
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\EventType whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\EventType whereName($value)
+ * @mixin \Eloquent
+ */
 class EventType extends Model
 {
     use LogsActivity;