+++ resources/views/emails/booking/changed.blade.php
@@ -3,7 +3,7 @@
 
 Dear **{{ $booking->user->full_name }}**,
 
-Your booking for the **{{ $booking->event->name }}** has been amended, please review the changes below:
+Your booking for the **{{ $booking->event->name }}** event has been amended, please review the changes below:
 
 @component('mail::table')
 |  |  |  |