+++ routes/web.php
@@ -13,8 +13,8 @@
 
 Route::get('/', 'HomeController')->name('home');
 
-Route::get('/login/{booking?}', 'Auth\LoginController@login')->name('login');
-Route::get('/validateLogin', 'Auth\LoginController@validateLogin');
+Route::get('/login', 'Auth\LoginController@login')->name('login');
+Route::get('/validate-login', 'Auth\LoginController@validateLogin');
 Route::get('/logout', 'Auth\LoginController@logout')->name('logout');
 
 // Admin routes